Add files via upload
In this case study I analyzed the correlation between exit velocity and slugging percentage versus the correlation between launch angle and slugging percentage. The goal of the analysis was to determine which of the two variables, when isolated as single indicators of success, correlates better with success. I have included the rmd file and a pdf file of the R markdown of my analysis.